separated profile and user
profile and user functions separated into own folders communicating with their respective collections
