+++ microsvcs/user/profile/app/controller/profile_con.js
@@ -1,7 +1,5 @@
 // Import Database Functions
-const QUERIES = require("../model/user.js")
-
-// LOCAL FUNCTIONS --------------------------------------------------------
+const QUERIES = require("../model/profile.js")
 
 // FUNCTION EXPORTS --------------------------------------------------------
 /**
@@ -25,20 +23,15 @@ exports.getProfileByUID = async (req, res) => {
     };
 };
 
-/**
- * Responds with a single document containing the
- * requested post
- * 
- * Expecting a JSON object from POST body.
- * @param {String} uid - unique uid field
- */
-exports.createNewUser = async (req, res) => {
+//TODO: Update Profile
+exports.updateProfileInfo = async (req, res) => {
     try {
-        const uid = req.body.email;
-        const name = req.body.name;
-
-        const result = await QUERIES.createNewUser(uid, name)
-        res.status(200).send(result);
+        const uid = req.params.uid;
+        const json_body = req.body
+        const result = await QUERIES.updateOneByUID(uid, json_body)
+        if (result.modifiedCount) {
+            res.status(200).send({ message: "success" });
+        }
     } catch (error) {
         if (error.message == "Server Error Occurred") {
             res.status(500).send({ message: error.message });
@@ -47,5 +40,3 @@ exports.createNewUser = async (req, res) => {
         }
     };
 };
-
-
