+++ microsvcs/user/profile/app/model/user.js
@@ -1,27 +1,16 @@
 const database = require('./util/init_db');
 
+
 //Define User  Collection
 const USER_COLLECTION = database.collection(process.env.USER_COLLECTION);
 
-//Define Profile Collection
-const PROFILE_COLLECTION = database.collection(process.env.PROFILE_COLLECTION);
-
 /**
- * Query the defined collection for the User/Scenario microservice.
+ * Insertone user into MongoDB
  * 
  * @param {String} uid - unique uid field
- * @return profiling data for the specified user
+ * @param {String} name - name
+ * @return success message or error thrown ie duplicate email entry
  */
-exports.findByUID = async (uid) => {
-    try {
-        let result = await PROFILE_COLLECTION.findOne({ uid: { $eq: uid } });
-        return result;
-    } catch (error) {
-        throw new Error("Server Error Occurred")
-    }
-}
-
-//TODO: POST new User
 exports.createNewUser = async (uid, name) => {
     try {
         let result = await USER_COLLECTION.insertOne({
